fix(trade-engine): price markets against the real horizon, not a rounded-up day
PolymarketScanner._horizon_days returned math.ceil((endDate - now) / 86400),
a verbatim port of the n8n node. The rounding was not conservative, it was
wrong in one fixed direction: a shorter real horizon rounded UP makes the
simulator diffuse the price for longer than the market has, which overstates
the probability of reaching the target, which inflates the edge, which
maximises position size.

On 2026-08-31 a 3,557.9-second market (0.0412d) was priced as a full day.
P went 0.0648 -> 0.4834 and edge +3.9pts -> +45.8pts against a 7-point floor,
so position e09b82fe fired at the $10 cap and lost $10.69.

Horizons are fractional end to end now, and the simulator derives dt from the
horizon (dt = horizon/steps, drift by dt, diffusion by sqrt(dt)) instead of
hardwiring dt = 1.0 and using the horizon as its own step count.

Two separate minimums, deliberately not conflated:

  MIN_HORIZON_TRADEABLE_DAYS = 1.0 is the control. Correct arithmetic does not
  make daily_sigma (21 daily closes) valid intraday, so sub-day markets are
  refused outright. Enforced twice and independently: the scanner drops them
  before simulation, executor GATE 7 re-checks against live state before the
  order. Env can raise it, never lower it.

  MIN_HORIZON_MODEL_DAYS = 0.01 is numerical safety only. It stops T = 0 and a
  hard 0.0000/1.0000 out of the simulator; it sits inside the degenerate region
  by design and is documented as not being a validity threshold.

A SECOND, independent truncation is also fixed: monte_carlo /simulate did
int(horizon), so a scanner-only fix would have made things worse. int(0.0412)
is 0, giving steps = 0 and a (paths x 0) array. np.any over an empty axis is
all-False, so every touch_* market would have priced at a silent P = 0.0 and
been persisted as real data; close_* markets fail loudly on paths[:, -1].

STEPS_PER_DAY is pinned at 1, the density that has always been in force.
Raising it is a sizing decision, not a tuning knob: it lifts every touch_*
probability and recruits markets from below the 7-point floor. Tracked
separately with the measurements.

Scope:
- scanner: _horizon_days returns float, tradeable-horizon refusal, 30.0 default
- simulation.py: kernel split out of monte_carlo.py so the maths is testable
  with numpy alone (importing monte_carlo needs flask + yfinance + scipy, which
  is why that module had no tests at all)
- models: ScannerCandidate.horizon_days int -> float. Left as int, pydantic v2
  raises int_from_float and takes down the whole scan, not one market
- executor: GATE 7, fails closed on missing or non-finite
- approval/analyst: format horizon as .2f for the human and the prompt
- persist horizon_days_model, steps and steps_per_day so a stored simulation
  records its own discretisation
- numpy into requirements.txt so the new tests run in CI

Trading stays disabled. This changes nothing about trading_config, and GATE 1
still re-reads the brake live on every execution.

+++ src/trade_engine/models.py
@@ -182,7 +182,11 @@ class MonteCarloResponse(BaseModel):
     current_price: float
     target: float
     asset: str
-    horizon_days: int
+    # Fractional. See ScannerCandidate.horizon_days for why this must not be
+    # int. NOTE this model is currently unused, nothing constructs it, but it
+    # is kept in step with the wire format rather than left as a stale shape
+    # that would reject a real worker response if it were ever wired up.
+    horizon_days: float
     market_type: str
     question: Optional[str] = None
     simulations: Optional[int] = None
@@ -225,7 +229,13 @@ class ScannerCandidate(BaseModel):
     sim_probability: float
     market_probability: float
     volume: float
-    horizon_days: int
+    # Fractional days to resolution, NOT a whole-day count. This annotation is
+    # load-bearing: pydantic v2 rejects a float with a fractional part for an
+    # int field (`int_from_float`), and _to_candidate below runs for every
+    # high-edge AND no-edge row, so leaving this as int turns the horizon fix
+    # into a ValidationError that takes down the whole scan rather than one
+    # market. Verified against pydantic 2.12.4.
+    horizon_days: float
     market_url: str
     amount_usdc: float
